Safari-609.1.11.1/LayoutTests/editing/selection/select-out-of-floated-non-editable-02.html

<!DOCTYPE html>
<html>
<body>
<div id="container"><div>block-level text before floats</div><div id="target" style="float: left; margin-right: 20px">first float*</div><div style="float: left; margin-right: 20px">second float*</div>inline-level text after the floats</div>
<p style="padding-top: 1em;">
To test manually, drag from the middle of the 'first float' div to the
right, just before the 'second float' div. The selection should go to the end
of the 'first float' div and not jump to the beginning.
</p>
<pre id="console"></pre>
<script src="../../resources/dump-as-markup.js"></script>
<script src="resources/select-out-of-floated-non-editable.js"></script>
<script>
runTest("float*", 20);
</script>
</body>
</html>
