Safari-609.1.11.1/LayoutTests/editing/selection/select-out-of-floated-non-editable-02-expected.txt

block-level text before floats
first float*second float*inline-level text after the floats
To test manually, drag from the middle of the 'first float' div to the right, just before the 'second float' div. The selection should go to the end of the 'first float' div and not jump to the beginning.

SUCCESS: Selected test is 'float*' as expected
| <div>
|   "block-level text before floats"
| <div>
|   id="target"
|   style="float: left; margin-right: 20px"
|   "first <#selection-anchor>float*"
| <div>
|   style="float: left; margin-right: 20px"
|   <#selection-focus>
|   "second float*"
| "inline-level text after the floats"
